redesign document interface and convert msg to json
OPEN_DOCUMENT_TREE is too buggy for native because:

1) The interface gives a uri but no mechanism exists to
   convert the uri to a directory path and the actual
   directory path may vary across Android devices.
2) The C opendir function cannot read files in the
   requested directory unless the user previously
   granted access. It doens't seem possible for a native
   app to manage a collection of files unless this worked.
3) The directory must lie under the "Documents" subdirectory
   on some versions of Android but this isn't enforced by
   the Android file picker.

The alternate method implemented in this patch is also
not ideal because it is uncommon to handle fds direclty
so significant changes were needed to redesign the code.

When the app closes, access to the file also goes away
until reopened via the Android file picker. This may be
solvable in the future but it's a clunky design. Perhaps
if there were an easy mechanism to convert a uri to a
file path then this could be easier to manage.